fix: Migrate unittest based test cases to pytest
